games/梦幻西游手游/scripts/requirements.txt
# 梦幻西游手游脚本 - 依赖
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
ppadb>=0.1.0
uiautomator2>=2.17.0
easyocr>=1.7.0
PyYAML>=6.0
requests>=2.31.0
